requirements.txt
django==2.1.1
pymongo==3.7.1
mongoengine==0.15.3
django-debug-toolbar
djongo
djangorestframework==3.8.2
Pillow==5.2.0
djangorestframework-jwt==1.11.0
redis==2.10.6
django_redis==4.9.0
requests==2.19.1
ujson
social-auth-app-django==2.1.0
unittest2==1.1.0
httpretty==0.9.5
django-filter==2.0.0
drf-extensions
django-cors-headers
coreapi
coreapi-cli
django-colorfield==0.1.15